+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## ????
+## 0.0.2 : 2023-12-18
 - Default headers: connection, accept, user-agent
 - Can provide options on each verb method
 - Support HTTP requests without curl, using file_get_contents()